upgrade parent and core versions

+++ src/main/java/io/github/davemeier82/homeautomation/zigbee2mqtt/device/property/Zigbee2MqttAlarm.java
@@ -28,7 +28,7 @@ public class Zigbee2MqttAlarm extends AbstractAlarm {
   private final String topic;
   private final MqttClient mqttClient;
 
-  public Zigbee2MqttAlarm(long id,
+  public Zigbee2MqttAlarm(int id,
                           Device device,
                           String topic,
                           EventPublisher eventPublisher,
